requirements.txt
grpcio
grpcio-tools
pika~=1.3.2
google~=3.0.0
redis~=5.0.3
<<<<<<< HEAD
PyYAML~=6.0.1
tabulate~=0.9.0
=======
pip~=24.0
setuptools~=69.5.1
>>>>>>> revisionM
